pyproject.toml
[project]
name = "tessera-ad-infinitum-generator"
version = "0.1.0"
dependencies = [
    "numpy",
    "pillow",
    "vnoise",
    "scipy",
    "pytest",
    "setuptools",
    "boto3",
]
